// packages/components/button/src/Button/Button.types.ts
export type ButtonType = "primary" | "secondary" | "tertiary";

export type ButtonInputType = "button" | "submit" | "reset";

export type ButtonColor =
  | "primary"
  | "positive"
  | "negative"
  | "inverted"
  | "on-primary-color"
  | "on-inverted-background"
  | "brand"
  | "fixed-light"
  | "fixed-dark";

export type ButtonSize = "xxs" | "xs" | "small" | "medium" | "large";